% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkCatalog}
\alias{checkCatalog}
\title{Check values in catalog}
\usage{
checkCatalog(catalog, site)
}
\arguments{
\item{catalog}{A dataframe describing the catalog}

\item{site}{Source of the data (usually a website)}
}
\value{
Dataframe of the catalog
}
\description{
Process multiple checks on the table.
}
