% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkDetails}
\alias{checkDetails}
\title{Check values in order details}
\usage{
checkDetails(details)
}
\arguments{
\item{details}{A dataframe describing the order details}
}
\value{
Dataframe of the order details
}
\description{
Process multiple checks on the table.
}
