% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkIfOneSiteMissing}
\alias{checkIfOneSiteMissing}
\title{Check files of one site}
\usage{
checkIfOneSiteMissing(missing)
}
\arguments{
\item{missing}{A dataframe describing missing sources}
}
\description{
Identify the files missing some sites.
}
