% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkMissingSites}
\alias{checkMissingSites}
\title{Check all files}
\usage{
checkMissingSites(files)
}
\arguments{
\item{files}{Dataframe describing files to load}
}
\value{
Dataframe describing files to load
}
\description{
Identify the files missing some sites.
}
