% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkOrders}
\alias{checkOrders}
\title{Check values in orders}
\usage{
checkOrders(orders, site)
}
\arguments{
\item{orders}{A dataframe describing the orders}

\item{site}{Source of the data (usually a website)}
}
\value{
Dataframe of the orders
}
\description{
Process multiple checks on the table.
}
