% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportOutput}
\alias{exportOutput}
\title{Export month report}
\usage{
exportOutput(df)
}
\arguments{
\item{df}{Dataframe to export}
}
\value{
Dataframe of the month report
}
\description{
Export dataframe to a CSV file
}
