% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parameters.R
\name{filterCompleteSites}
\alias{filterCompleteSites}
\title{Filter complete sites}
\usage{
filterCompleteSites(files)
}
\arguments{
\item{files}{Dataframe describing files to load}
}
\value{
Dataframe of files to load
}
\description{
Get files description of complete sites.
}
