% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{getColumnNames}
\alias{getColumnNames}
\title{Get column names}
\usage{
getColumnNames(hasIds, table)
}
\arguments{
\item{hasIds}{A boolean identifying a table with a column id}

\item{table}{The table type}
}
\value{
Vector of column names
}
\description{
Identify column names based on the table type.
}
