% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{getDelimiter}
\alias{getDelimiter}
\title{Get delimiter}
\usage{
getDelimiter(delim)
}
\arguments{
\item{delim}{Delimiter code defined in parameters}
}
\value{
Character used to delimit fields in a file
}
\description{
Identify delimiter used in the CSV file
}
