% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parameters.R
\name{getFiles}
\alias{getFiles}
\title{Filter complete sites}
\usage{
getFiles()
}
\value{
Dataframe of files to load
}
\description{
Get all files description and add a field showing
if the file exists or not.
}
