% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{getFinalTable}
\alias{getFinalTable}
\title{Merge tables of a site}
\usage{
getFinalTable(orders, details, catalog, site)
}
\arguments{
\item{orders}{Dataframe describing orders}

\item{details}{Dataframe describing order details}

\item{catalog}{Dataframe describing products catalog}

\item{site}{Source of the data (usually a website)}
}
\value{
Dataframe of the site
}
\description{
Take the 3 files of a site - catalog, orders and details -
and apply a filter, a join and an agregation.
}
