% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parameters.R
\name{getSites}
\alias{getSites}
\title{Get complete sites}
\usage{
getSites(files)
}
\arguments{
\item{files}{Dataframe describing files to load}
}
\value{
Dataframe of files from site that are fully defined
}
\description{
Get complete sites
}
