% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-policy.R
\name{incErrorsPolicy}
\alias{incErrorsPolicy}
\title{Increment errors found by policy}
\usage{
incErrorsPolicy(policy)
}
\arguments{
\item{policy}{The id of the policy to increment}
}
\description{
When an error is found, this function increments the
number of errors for a defined policy
}
