% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parameters.R
\name{loadParameters}
\alias{loadParameters}
\title{Load parameters of sites}
\usage{
loadParameters(filename)
}
\arguments{
\item{filename}{Name of the file describing the files format}
}
\value{
Dataframe of sites
}
\description{
Load parameters of sites
}
