% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadSite}
\alias{loadSite}
\title{Load a site}
\usage{
loadSite(files, site)
}
\arguments{
\item{files}{Dataframe describing files to load}

\item{site}{Source of the data (usually a website)}
}
\value{
Dataframe of the site
}
\description{
Load the 3 files of a site: catalog, orders and details.
}
