% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadSites}
\alias{loadSites}
\title{Load all sites}
\usage{
loadSites(files)
}
\arguments{
\item{files}{Dataframe describing files to load}
}
\value{
Dataframe of all sites
}
\description{
Load all sites described in parameters.
}
