% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{loadTable}
\alias{loadTable}
\title{Load CSV file}
\usage{
loadTable(files, site, table)
}
\arguments{
\item{files}{Dataframe describing files to load}

\item{site}{Source of the data (usually a website)}
}
\value{
Dataframe of the file content
}
\description{
The output has the right column names, types and units.
}
