% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{monthReport}
\alias{monthReport}
\title{Get month}
\usage{
monthReport()
}
\value{
Dataframe of the month report
}
\description{
Make the whole job of filtering, joining and aggregating data of all sites.
Before execution, it loads the parameters describe in the .env file,
checks the parameters and launch the whole process.
}
