% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{selectColumns}
\alias{selectColumns}
\title{Get column names (without id field)}
\usage{
selectColumns(table)
}
\arguments{
\item{table}{The table type}
}
\value{
Vector of column names
}
\description{
Identify column names based on the table type. Removed id fields if it exists.
}
