% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data.R
\name{checkDetails}
\alias{checkDetails}
\title{Order details
orderId, EAN and quantity are always defined for each record
Quantity are never under 1 or greater than 10
Quantity are usually between 1 and 5
Quantity are usually integer
details}
\usage{
checkDetails(details)
}
\description{
Order details
orderId, EAN and quantity are always defined for each record
Quantity are never under 1 or greater than 10
Quantity are usually between 1 and 5
Quantity are usually integer
details
}
